% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_abline}
\alias{add_abline}
\title{Wrapper to add lines to nested loop plot}
\usage{
add_abline(
  p,
  slope = 0,
  intercept = 0,
  position = "bottom",
  linetype = 3,
  size = 0.5,
  colour = "black",
  alpha = 1,
  ...
)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{position}{Position of the geometry layer, either "bottom" or "top". By default
lines are added below all other lines in the plot to not obscure them, 
but by setting position to "top", these lines can also be plotted 
above other lines.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
All parameters not explicitly described here are simply passed to 
\code{\link{geom_abline}}.
}
\details{
Simple wrapper for \code{\link{geom_abline}} to add  
lines to nested loop plot. Useful to display e.g. target values in an 
experimental study, e.g. zero prediction error or zero bias (when used
with slope = 0 and intercept = 0).
}
