% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_custom_theme}
\alias{add_custom_theme}
\title{Wrapper to customize theme of nested loop plot}
\usage{
add_custom_theme(p, ...)
}
\arguments{
\item{p}{\pkg{ggplot2} object, Output from \code{\link{nested_loop_plot}} function.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
All parameters are passed to \code{\link{theme}}.
}
