% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_geom_at_position}
\alias{add_geom_at_position}
\title{Adds \pkg{ggplot2} derived geometry at specified z-level position}
\usage{
add_geom_at_position(p, geom, position = "top")
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{geom}{\pkg{ggplot2} derived geometry which should be added to the plot object.}

\item{position}{Position of the geometry layer, either "bottom" or "top". By default
lines are added below all other layers in the plot to not obscure them, 
but by setting position to "top", these lines can also be plotted 
above other lines.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Adds \pkg{ggplot2} derived geometry at specified z-level position
}
\details{
Simple wrapper to add arbitrary geometries to nested loop plot.
}
