% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_lines}
\alias{add_lines}
\title{Wrapper to add lines to nested loop plot}
\usage{
add_lines(p, position = "bottom", ...)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{position}{Position of the geometry layer, either "bottom" or "top". By default
lines are added below all other layers in the plot to not obscure them, 
but by setting position to "top", these lines can also be plotted 
above other lines.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
All parameters not explicitly described here are simply passed to 
\code{\link{geom_line}}.
}
\details{
Simple wrapper for \code{\link{geom_line}} to add  
lines to nested loop plot.
}
