% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_points}
\alias{add_points}
\title{Wrapper to add points to nested loop plot}
\usage{
add_points(p, position = "top", ...)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{position}{Position of the geometry layer, either "bottom" or "top". By default
points are added above all other geometries in the plot to be unobscured, 
but by setting position to "top", these points can also be plotted 
below other geometries.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
All parameters not explicitly described here are simply passed to 
\code{\link{geom_point}}.
}
\details{
Simple wrapper for \code{\link{geom_point}} to add  
points to nested loop plot. Useful to e.g. indicate special values.
}
