% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_processing}
\alias{add_processing}
\title{Add post processing to \pkg{ggplot2} object}
\usage{
add_processing(p, processing = NULL)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{post_processing}{List with post processing wrappers to apply to the plot. Name of an entry
specifies the function to call, the entry itself contains named arguments
for the wrapper function which are passed to \code{do.call}.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Add post processing to \pkg{ggplot2} object
}
\details{
Uses \code{do.call} to call the functions. \code{p} is always passed first 
and does not need to be specified manually.
}
