% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_steps}
\alias{add_steps}
\title{Wrapper to add steps to nested loop plot}
\usage{
add_steps(p, position = "bottom", ...)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{position}{Position of the geometry layer, either "bottom" or "top". By default
steps are added below all other layers in the plot to not obscure them, 
but by setting position to "top", these steps can also be plotted 
above other lines.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
All parameters not explicitly described here are simply passed to 
\code{\link{geom_step}}.
}
\details{
Simple wrapper for \code{\link{geom_step}} to add  steps to nested loop plot.
}
