% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{add_text}
\alias{add_text}
\title{Wrapper to add text to nested loop plot}
\usage{
add_text(p, decimals = 2, mapping = NULL, data = NULL, position = "top", ...)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}} function.}

\item{decimals}{Number of decimals to be display if called with default mapping to display
labels for each plotted value.}

\item{mapping}{Aesthethics mapping specification for the text created by \code{\link{aes}}. 
This allows to use user-specified data as labels. Passed to 
\code{\link{geom_text}}. The mapping must specify the \code{label}
aesthetic.}

\item{data}{Specify other data source to use label data from. Must contain information
from the data.frame that is plotted and obtained from 
\code{\link{nested_loop_base_data}}. Passed to \code{\link{geom_text}}.}

\item{position}{Position of the geometry layer, either "bottom" or "top". By default
text is added above all other layers in the plot to be unobstructed by other
geometries, but by setting position to "bottom", text can also be plotted 
below other geometries.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
All parameters not explicitly described here are simply passed to 
\code{\link{geom_text}}.
}
\details{
Simple wrapper for \code{\link{geom_text}} to add  
text to nested loop plot. Useful to e.g. annotate special values.

The most simple, default usage is to add the values of the plots (rounded
to two decimals). The plotted values are accessed via the \code{y_coord} variable
of the data.frame returned from \code{\link{nested_loop_base_data}}.

More advanced usage and custom labeldata can be introduced to the plots
by using the modular functions of this package. See the demo vignette for
an example.
}
\examples{
\dontrun{
# add values with two decimals to plot 
p = add_text(p, aes = aes(labels = round(values, 2)))
}

}
