% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{adjust_ylim}
\alias{adjust_ylim}
\title{Adjust y-axis limits of \pkg{ggplot2} object}
\usage{
adjust_ylim(p, y_expand_mult = NULL, y_expand_add = NULL)
}
\arguments{
\item{y_expand_mult}{NULL, a named list or a vector of length 2. 
Used for adjustments to the display area similar
to what \code{\link{expand_scale}} does. The lower limit of display
will be expanded by a fraction of the plotting range as given by the first 
entry of the vector, the upper limit by a fraction according to the second
entry. Useful to adjust the y-axis when steps for design parameters are drawn
below the results. If a list, then the names of the list should correspond
to facet panels of the plot and the entries should be vectors of length two,
which apply the scale expansion only to the specified facets.
If NULL, then no adjustment is done.}

\item{y_expand_add}{NULL, a named list or a vector of length 2. 
Used for adjustments to the display area similar
to what \code{\link{expand_scale}} does. The lower limit of display 
will be changed by addition of the first entry, the upper limit by addition of 
the second entry. Specified in y-axis coordinates. Useful to adjust the 
y-axis when steps for design parameters are drawn below the results.
If a list, then the names of the list should correspond
to facet panels of the plot and the entries should be vectors of length two,
which apply the scale expansion only to the specified facets.
If NULL, then no adjustment is done.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Set y-axis limits in a way which is useful for the output from 
\code{\link{nested_loop_plot}}. Makes functionality of 
\code{\link{expand_scale}} available.
}
\details{
Often helpful to deal with clipping issue when e.g. text is not fully 
displayed within the plot. 

This function extends the axis range via additive or multiplicative 
factors.

Note that all numeric values are converted to absolute values (i.e. this 
function can not be used to shrink the y-axis).
}
