% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{expand_ylim}
\alias{expand_ylim}
\title{Adjust y-axis limits of \pkg{ggplot2} object}
\usage{
expand_ylim(p, y_expand_limits = 0)
}
\arguments{
\item{p}{\pkg{ggplot2} object, Output from \code{\link{nested_loop_plot}} function.}

\item{y_expand_limits}{A named list or a numeric vector. 
If a list, then the names of the list should correspond
to facet panels of the plot and the entries should be numeric vectors,
which apply the scale expansion only to the specified facets.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Set y-axis limits in a way which is useful for the output from 
\code{\link{nested_loop_plot}}. Makes functionality from
\code{\link{expand_scale}} available.
}
\details{
Often helpful to deal with clipping issue when e.g. text is not fully 
displayed within the plot. 

This function makes sure that any value in \code{y_expand_limits} is visible
in the plot.
}
