% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{label_both_custom}
\alias{label_both_custom}
\title{Facet labeller helper}
\usage{
label_both_custom(labels, multi_line = TRUE, sep = ": ")
}
\description{
Customized \code{label_both} from \pkg{ggplot2} 3.0.
}
\details{
Removes the suffix "_labels_" from the variable names passed to the 
labeller. This suffix is added during processing of the nested loop data.
}
