% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{nested_loop_paramsteps_data}
\alias{nested_loop_paramsteps_data}
\title{Generate data for facetted nested loop plots}
\usage{
nested_loop_paramsteps_data(
  plot_data,
  steps_y_base = 0,
  steps_y_height = 1,
  steps_y_shift = NULL
)
}
\arguments{
\item{plot_data}{Outout list from \code{\link{nested_loop_base_data}}.}

\item{steps_y_base}{Numeric. Maximum height of steps in y-axis units. I.e. if steps are 
increasing (due to \code{parameter_decreasing == FALSE}) this represents the
y-axis value of the uppermost step, if \code{parameter_decreasing == TRUE} in 
\code{plot_data} this is the y-axis value of the first step.}

\item{steps_y_height}{Numeric. Height of a single step in y-axis units. If a single numeric, 
the same height is used for all steps. If a vector, then the step heights
may vary for each layer (as defined by \code{steps} argument). Values
are cycled to have appropriate length.}

\item{steps_y_shift}{Numeric. Distance in y-axis units between step layers, i.e. distance 
between step drawn for different design parameters (if \code{steps} 
comprises more than one variable). As \code{steps_y_height}, this 
can be a vector to allow varying shift between layers. If NULL, an 
automated attempt is made to set the value to \code{0.25*steps_y_height}, but
this may need manual tweaking.}
}
\value{
Updates the list\code{plot_data}:

\enumerate{
\item Adds an entry for \code{stepdf}, which is a data.frame containing the 
values for the parameter steps to be plotted. 
\item Adds user input as entries to the list \code{input}: *steps_y_base, 
steps_y_shift, steps_y_height*.
}
}
\description{
This function generates coordinates for plotting steps for design 
parameters based on the data from \code{\link{nested_loop_base_data}}. 
All parameter definitions are the same as in \code{\link{nested_loop_plot}}.
}
