% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_loop.R
\name{nested_loop_paramsteps_plot}
\alias{nested_loop_paramsteps_plot}
\title{Adds step functions for design parameters to a basic facetted nested loop plot}
\usage{
nested_loop_paramsteps_plot(
  p,
  step_data,
  steps_names = NULL,
  steps_names_annotate = TRUE,
  steps_values_annotate = FALSE,
  steps_color = "#AAAAAA",
  steps_annotation_size = 5,
  steps_annotation_nudge = 0.2,
  steps_annotation_color = "#AAAAAA"
)
}
\arguments{
\item{p}{\pkg{ggplot2} object, output from \code{\link{nested_loop_plot}}.}

\item{steps_names}{NULL or character value of same length as \code{steps}. Specifies names of the 
design parameters which are used for steps.}

\item{steps_names_annotate}{Logical. Should steps drawn be annotated with names of corresponding
design parameters?}

\item{steps_values_annotate}{Logical. Should steps drawn be annotated with values of corresponding
design parameters? Only the first occurence of a new value is annotated
to avoid visual clutter.}

\item{steps_color}{Color specification for steps drawn.}

\item{steps_annotation_size}{Numeric. Size of annotations for steps. Likely needs tweaking.}

\item{steps_annotation_nudge}{Numeric. Fine-tune position of steps annotations in y-axis units. 
Often, the annotation is overlayed with the lines of the steps - this 
argument simply increases the distance between annotations and step lines,
similar to the \code{nudge_y} argument of \code{\link{geom_text}}.}

\item{steps_annotation_color}{Color specification of the step annotation text.}

\item{plot_data}{Output list from \code{\link{nested_loop_base_data}}.}
}
\value{
An updated \pkg{ggplot2} object. (Note that this function does NOT return a 
deep copy of the original plot but rather updates it.)
}
\description{
Adds steps for design parameters to a basic nested loop plots, i.e. output of
 \code{\link{nested_loop_base_plot}}.
 All parameter definitions are the same as in \code{\link{nested_loop_plot}}.
}
